add cloudformation templates for reports
add README for UserList
